Move: Move Entities directory to Domains directory
