+++ Domains/Entities/albumEntity.js
