+++ Domains/Entities/authenticationEntity.js
